pyproject.toml
[project]
name = "leetcodebuddy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "discord-py>=2.5.2",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "selenium>=4.34.2",
    "supabase>=2.17.0",
]
